+++ libs/libc/pthread/CMakeLists.txt
@@ -89,9 +89,15 @@ if(NOT CONFIG_DISABLE_PTHREAD)
     pthread_mutexattr_getrobust.c
     pthread_mutexattr_setprioceiling.c
     pthread_mutexattr_getprioceiling.c
+    pthread_mutex.c
+    pthread_mutex_destroy.c
+    pthread_mutex_getprioceiling.c
+    pthread_mutex_init.c
     pthread_mutex_lock.c
     pthread_mutex_setprioceiling.c
-    pthread_mutex_getprioceiling.c
+    pthread_mutex_timedlock.c
+    pthread_mutex_trylock.c
+    pthread_mutex_unlock.c
     pthread_once.c
     pthread_yield.c
     pthread_atfork.c
@@ -109,6 +115,10 @@ if(NOT CONFIG_DISABLE_PTHREAD)
     pthread_self.c
     pthread_gettid_np.c
     pthread_concurrency.c)
+
+  if(NOT CONFIG_PTHREAD_MUTEX_UNSAFE)
+    list(APPEND SRCS pthread_mutex_consistent.c)
+  endif()
 
   if(CONFIG_SMP)
     list(APPEND SRCS pthread_attr_getaffinity.c pthread_attr_setaffinity.c)
