+++ libs/libc/pthread/Make.defs
@@ -55,6 +55,8 @@ CSRCS += pthread_mutexattr_setprotocol.c pthread_mutexattr_getprotocol.c
 CSRCS += pthread_mutexattr_settype.c pthread_mutexattr_gettype.c
 CSRCS += pthread_mutexattr_setrobust.c pthread_mutexattr_getrobust.c
 CSRCS += pthread_mutexattr_setprioceiling.c pthread_mutexattr_getprioceiling.c
+CSRCS += pthread_mutex_destroy.c pthread_mutex_init.c pthread_mutex_trylock.c
+CSRCS += pthread_mutex_timedlock.c pthread_mutex.c pthread_mutex_unlock.c
 CSRCS += pthread_mutex_lock.c
 CSRCS += pthread_mutex_setprioceiling.c pthread_mutex_getprioceiling.c
 CSRCS += pthread_once.c pthread_yield.c pthread_atfork.c
@@ -68,6 +70,10 @@ CSRCS += pthread_concurrency.c
 
 ifeq ($(CONFIG_SMP),y)
 CSRCS += pthread_attr_getaffinity.c pthread_attr_setaffinity.c
+endif
+
+ifneq ($(CONFIG_PTHREAD_MUTEX_UNSAFE),y)
+CSRCS += pthread_mutex_consistent.c
 endif
 
 ifeq ($(CONFIG_PTHREAD_SPINLOCKS),y)
