+++ libs/libc/pthread/pthread_mutex_consistent.c
@@ -1,5 +1,5 @@
 /****************************************************************************
- * sched/pthread/pthread_mutexconsistent.c
+ * libs/libc/pthread/pthread_mutex_consistent.c
  *
  * SPDX-License-Identifier: Apache-2.0
  *
@@ -33,8 +33,7 @@
 #include <errno.h>
 
 #include <nuttx/semaphore.h>
-
-#include "pthread/pthread.h"
+#include <nuttx/pthread.h>
 
 /****************************************************************************
  * Public Functions
@@ -102,7 +101,7 @@ int pthread_mutex_consistent(FAR pthread_mutex_t *mutex)
            * nxsched_get_tcb() does.
            */
 
-          if (nxsched_get_tcb(pid) == NULL)
+          if (pthread_kill(pid, 0) != 0)
             {
               /* Reset the semaphore.  This has the same affect as if the
                * dead task had called pthread_mutex_unlock().
