+++ sched/task/task_fork.c
@@ -151,10 +151,6 @@ FAR struct tcb_s *nxtask_setup_fork(start_t retaddr)
 
   nxtask_joininit(child);
 
-#if !defined(CONFIG_DISABLE_PTHREAD) && !defined(CONFIG_PTHREAD_MUTEX_UNSAFE)
-  spin_lock_init(&child->mhead_lock);
-#endif
-
   /* Allocate a new task group with the same privileges as the parent */
 
   ret = group_allocate(child, ttype);
